pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "strads-alert-agent"
version = "0.1.0"
description = "Intelligent Alert Decision Agent - Multi-agent pipeline for alert triage and decision making"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "strands-agents>=1.21.0",
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "requests>=2.31.0",
    "neo4j>=5.0.0",
    "chromadb>=0.4.0",
    "prometheus-client>=0.17.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "prometheus-api-client>=0.7.0",
    "jinja2>=3.1.6",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "azure-ai-inference>=1.0.0b9",
    "qdrant-client>=1.7.0",
    "strands-agents-tools>=0.2.19",
    "slack-bolt>=1.18.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.0.280",
    "mypy>=1.4.0",
    "pytest-cov>=4.1.0",
]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.poetry.source]]
name = "pypi-bb-dev-local"
url = "https://binarios.intranet.bb.com.br/artifactory/api/pypi/pypi-bb-dev-local/simple"
priority = "primary"

